chore(VLE): Convert Node Controller to Angular
#531
