+++ src/assets/wise5/authoringTool/main/mainAuthoring.html
@@ -2,7 +2,7 @@
   <div class="l-constrained" layout="column">
     <md-content
         class="node-content md-whiteframe-1dp"
-        style="border-color: {{nodeController.nodeStatus.icon.color}}; position: relative;"
+        style="position: relative;"
         layout="row" layout-wrap>
       <style>
         .projectItem:hover {
