+++ src/assets/wise5/authoringTool/project/projectAuthoring.html
@@ -1,7 +1,7 @@
 <div id="top" class="view-content view-content--with-sidemenu">
   <div class="l-constrained" layout="column">
     <md-content class="node-content md-whiteframe-1dp"
-        style="padding: 0 16px; border-color: {{::nodeController.nodeStatus.icon.color}}; position: relative;"
+        style="padding: 0 16px; position: relative;"
         layout="row" layout-wrap>
       <div style="margin-bottom: 20px; width: 100%">
         <div style='text-align: right; background-color: white; position: sticky; top: 0px; z-index: 2;'>
