--- src/assets/wise5/themes/default/node/node.html
@@ -1,48 +1,0 @@
-<div class="l-constrained">
-    <div ng-if="::nodeController.isEndedAndLocked" class="notice node-notice" layout="row" layout-align="center start">
-        <span><md-icon class="warn"> lock </md-icon>&nbsp;</span>
-        <span class="warn md-body-2">{{ ::nodeController.endedAndLockedMessage }}</span>
-      </div>
-    <md-content class="node-content md-whiteframe-1dp" style="border-color: {{nodeController.nodeStatus.icon.color}}; position: relative;" layout="row" layout-wrap>
-        <div ng-if="::nodeController.isShowNodeRubric()" class="node-content__rubric" layout="row" layout-align="center center">
-            <help-icon id="nodeRubric_{{nodeController.node.id}}"
-                rubric-id="{{nodeController.node.id}}"
-                [pulse]="!nodeController.rubricTour.steps[0].viewed"
-                label="{{ 'STEP_INFO' | translate}}"
-                icon="info"></help-icon>
-        </div>
-        <div id="component_{{component.id}}" class="component"
-             ng-repeat='component in ::nodeController.getComponents() track by component.id'
-             flex="100" flex-gt-sm="{{component.componentWidth ? component.componentWidth : 100}}">
-            <div ng-if="::nodeController.isShowComponentRubric(component)" class="component__rubric">
-                <help-icon id="rubric_{{component.id}}"
-                    rubric-id="{{component.id}}"
-                    [pulse]="!nodeController.rubricTour.steps[$index].viewed"
-                    label="{{ 'TEACHING_TIPS' | translate }}"
-                    icon="info"></help-icon>
-            </div>
-            <component node-id='{{::nodeController.node.id}}'
-                       component-id='{{::component.id}}'
-                       [component-state]='nodeController.getComponentStateByComponentId(component.id)'
-                       [workgroup-id]='nodeController.workgroupId'></component>
-        </div>
-        <div class="node-content__actions" layout="row" layout-align="start center" flex="100">
-            <md-button class="md-raised md-primary"
-                   ng-click='nodeController.saveButtonClicked()'
-                   ng-disabled='nodeController.isDisabled || !nodeController.dirtyComponentIds.length'
-                   ng-if='::nodeController.node.showSaveButton'
-                   aria-label="{{ ::'save' | translate }}"
-                   translate="save"></md-button>
-             <md-button class="md-raised md-primary"
-                  ng-click='nodeController.submitButtonClicked()'
-                  ng-disabled='nodeController.isDisabled || !nodeController.dirtySubmitComponentIds.length'
-                  ng-if='::nodeController.node.showSubmitButton'
-                  aria-label="{{ ::'submit' | translate }}"
-                  translate="submit"></md-button>
-            <span ng-if="nodeController.saveMessage.text && (nodeController.node.showSaveButton || nodeController.node.showSubmitButton)"
-                  class="node-content__actions__info md-caption">
-                  {{nodeController.saveMessage.text}} <span class="node-content__actions__more"><md-tooltip md-direction="top">{{ nodeController.saveMessage.time | amDateFormat:'ddd, MMM D YYYY, h:mm a' }}</md-tooltip><span am-time-ago="nodeController.saveMessage.time"></span></span>
-            </span>
-        </div>
-    </md-content>
-</div>
