+++ src/assets/wise5/themes/default/style/modules/_help.scss
@@ -43,102 +43,3 @@
     border-bottom-left-radius: $card-border-radius;
     border-bottom-right-radius: $card-border-radius;
 }
-
-div.hopscotch-bubble {
-    border-radius: $card-border-radius;
-    //border: 2px solid color('gray-darker');
-    border: 0 none;
-    font-family: Roboto, "Helvetica Neue", sans-serif;
-    font-size: rem(1.4);
-    z-index: 6;
-
-    .hopscotch-bubble-arrow-container {
-        position: absolute;
-        width: 20px;
-        height: 20px;
-    }
-
-    .hopscotch-bubble-arrow-container {
-        &.up {
-            top: 0;
-            left: 12px;
-
-            .hopscotch-bubble-arrow {
-                border-bottom: 10px solid color('info');
-                border-left: 10px solid transparent;
-                border-right: 10px solid transparent;
-                position: relative;
-                top: -10px;
-            }
-
-            .hopscotch-bubble-arrow-border {
-                border: 0 none;
-                /*border-bottom: 12px solid color('gray-darker');
-                border-left: 12px solid transparent;
-                border-right: 12px solid transparent;*/
-            }
-        }
-
-        &.down {
-            bottom: -34px;
-            left: 12px;
-
-            .hopscotch-bubble-arrow {
-                border-top: 10px solid color('info');
-                border-left: 10px solid transparent;
-                border-right: 10px solid transparent;
-                position: relative;
-                top: -10px;
-            }
-
-            .hopscotch-bubble-arrow-border {
-                border: 0 none;
-                /*border-top: 12px solid color('gray-darker');
-                border-left: 12px solid transparent;
-                border-right: 12px solid transparent;*/
-            }
-        }
-
-        &.left {
-            top: 12px;
-            left: -10px;
-
-            .hopscotch-bubble-arrow {
-                border-right: 10px solid color('info');
-                border-bottom: 10px solid transparent;
-                border-top: 10px solid transparent;
-                position: relative;
-                left: 0;
-                top: -10px;
-            }
-
-            .hopscotch-bubble-arrow-border {
-                border: 0 none;
-                /*border-right: 12px solid color('gray-darker');
-                border-bottom: 12px solid transparent;
-                border-top: 12px solid transparent;*/
-            }
-        }
-
-        &.right {
-            top: 12px;
-            right: -30px;
-
-            .hopscotch-bubble-arrow {
-                border-left: 10px solid color('info');
-                border-bottom: 10px solid transparent;
-                border-top: 10px solid transparent;
-                position: relative;
-                right: 0;
-                top: -10px;
-            }
-
-            .hopscotch-bubble-arrow-border {
-                border: 0 none;
-                /*border-right: 12px solid color('gray-darker');
-                border-bottom: 12px solid transparent;
-                border-top: 12px solid transparent;*/
-            }
-        }
-    }
-}
