+++ src/assets/wise5/themes/default/vle.html
@@ -1,4 +1,3 @@
-<link rel="stylesheet" href="/assets/wise5/lib/hopscotch/dist/css/hopscotch.min.css" />
 <link rel="stylesheet" href="/assets/wise5/themes/default/style/angular-material.css" />
 <link rel="stylesheet" href="/assets/wise5/themes/default/style/vle.css" />
 <style ng-bind-html="::vleController.projectStyle"></style>
