+++ src/assets/wise5/vle/node/index.html
@@ -1,1 +1,1 @@
-<div ng-include src="::vleController.themePath + '/node/node.html'" layout="column"></div>
+<node></node>
