+++ src/assets/wise5/vle/node/node.component.html
@@ -1,0 +1,81 @@
+<div class="l-constrained">
+  <div *ngIf="isEndedAndLocked"
+      class="notice node-notice"
+      fxLayout="row"
+      fxLayoutAlign="center start">
+    <span><mat-icon class="warn">lock</mat-icon>&nbsp;</span>
+    <span class="warn md-body-2">{{ endedAndLockedMessage }}</span>
+  </div>
+  <div class="node-content md-whiteframe-1dp"
+      style="border-color: {{nodeStatus?.icon?.color}}; position: relative;"
+      fxLayout="row wrap">
+    <div *ngIf="isShowNodeRubric()"
+        class="node-content__rubric"
+        fxLayout="row"
+        fxLayoutAlign="center center">
+      <help-icon
+          [content]="replaceAssetPaths(rubric)"
+          [pulse]="idToIsPulsing[node.id]"
+          (click)="helpIconClicked(node.id)"
+          label="Step Info"
+          i18n-label
+          icon="info">
+      </help-icon>
+    </div>
+    <div *ngFor="let component of getComponents(); index as componentIndex"
+        class="component"
+        fxFlex
+        fxFlex.gt-sm="{{component.componentWidth ? component.componentWidth : 100}}">
+      <div *ngIf="isShowComponentRubric(component)" class="component__rubric">
+        <help-icon
+            [content]="replaceAssetPaths(component.rubric)"
+            [pulse]="idToIsPulsing[component.id]"
+            (click)="helpIconClicked(component.id)"
+            label="Teaching Tips"
+            i18n-label
+            icon="info">
+        </help-icon>
+      </div>
+      <component
+          [nodeId]="node.id"
+          [componentId]="component.id"
+          [componentState]="getComponentStateByComponentId(component.id)"
+          [workgroupId]="workgroupId">
+      </component>
+    </div>
+    <div class="node-content__actions"
+        fxLayout="row"
+        fxLayoutAlign="start center"
+        fxLayoutGap="12px">
+      <button *ngIf="node.showSaveButton"
+          mat-button
+          class="mat-primary mat-raised-button"
+          (click)="saveButtonClicked()"
+          [disabled]="isDisabled || !dirtyComponentIds.length"
+          aria-label="Save"
+          i18n-aria-label
+          i18n>
+        Save
+      </button>
+      <button *ngIf="node.showSubmitButton"
+          mat-button
+          class="mat-primary mat-raised-button"
+          (click)="submitButtonClicked()"
+          [disabled]="isDisabled || !dirtySubmitComponentIds.length"
+          aria-label="Submit"
+          i18n-aria-label
+          i18n>
+        Submit
+      </button>
+      <span *ngIf="saveMessage.text && (node.showSaveButton || node.showSubmitButton)"
+          class="node-content__actions__info md-caption">
+        {{ saveMessage.text }}
+        <span class="node-content__actions__more"
+            matTooltip="{{ saveMessage.time | amDateFormat:'ddd, MMM D YYYY, h:mm a' }}"
+            matTooltipPosition="above">
+          {{ saveMessage.time | amTimeAgo }}
+        </span>
+      </span>
+    </div>
+  </div>
+</div>
