+++ src/assets/wise5/vle/node/node.component.scss
@@ -1,0 +1,3 @@
+.mat-button {
+  min-width: 88px;
+}
